fix: improve furigana generation speed
